The syntax annotations are added to the words. Due to possible ambiguity there are multiple possible variants.
Not sure if this is the best way, I would ideally like to see syntax layer stuff in the analysis themselves. But as syntax analyzer has its own disambiguator, I am not sure at the moment how to solve this so that it would be future-proof. I'll let the pratice decide.